Create redux store and re-organize folders
