+++ src/App.js
@@ -1,10 +1,13 @@
 import './App.css';
 import React, {useState, useEffect} from "react"
-import Header from "./header/Header.jsx"
-import Body from "./body/Body.jsx"
-import Footer from "./footer/Footer.jsx"
-import {FetchData} from "./body/RequestPictures.jsx"
-import { v4 as uuidv4 } from 'uuid';
+import Header from "./components/header/Header.jsx"
+import Body from "./components/body/Body.jsx"
+import Footer from "./components/footer/Footer.jsx"
+import {FetchData} from "./components/body/RequestPictures.jsx"
+import { v4 as uuidv4 } from 'uuid'
+import Modal from "./components/modal/Modal.jsx"
+import store from "./store.js"
+import { Provider } from "react-redux"
 
 
 const _ = require("lodash")
@@ -25,7 +28,6 @@ function App() {
   useEffect(() => {
     if (dataArray.length === 0 || dataArray.length < 12) {
       FetchData(12, {setDataArray, dataArray})
-      console.log(dataArray)
     }
   }, [dataArray]);
   
@@ -47,30 +49,32 @@ function App() {
   
     
     return (
-      
-    <div className="App">
-      <header className="App-header">
-          <Header />
-      </header>
+    <Provider store={store}>
+      <div className="App">
+        <header className="App-header">
+            <Header />
+        </header>
 
 
-      <div className="App-body">
-        {
-          mapData.map(card => {
-            return <Body key={uuidv4()} dataArray={card} bool={_.isEmpty(card) ? false : true} thumbs={thumbnailURL}/>  
-          })
-        } 
-        <button className="loader" onClick={handleClick}>
-          {"Load More"}
-        </button>
+        <div className="App-body">
+          {
+            mapData.map(card => {
+              return <Body key={uuidv4()} dataArray={card} bool={_.isEmpty(card) ? false : true} thumbs={thumbnailURL}/>  
+            })
+          } 
+          <button className="loader" onClick={handleClick}>
+            {"Load More"}
+          </button>
+        </div>
+
+
+        <div className="App-footer">
+          <Footer/>
+        </div>
+
+        <Modal />
       </div>
-
-
-      <div className="App-footer">
-        <Footer />
-      </div>
-    </div>
-
+    </Provider>
   );
 }
