+++ src/components/modal/Modal.jsx
@@ -1,0 +1,18 @@
+import "./modal.css"
+import React, {useState} from "react"
+
+
+
+
+const Modal = () => {
+  const [modal, setModal] = useState(true);
+
+  
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Modal
