+++ src/components/modal/modal.css
@@ -1,0 +1,3 @@
+.bg-gray-modal {
+  background-color: gray !important;
+}
