+++ src/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from '@reduxjs/toolkit'
+import modalSlice from "./features/modal/modalSlice"
+
+export default configureStore({
+  reducer: {
+    toggle: modalSlice,
+  },
+})
